+++ Plain.h
@@ -24,7 +24,10 @@ class CPlain
         size_t NumLegs()               { return m_config.m_num_legs; }
         size_t LegStart(size_t index)  { return m_config.m_start[index]; }
         size_t LegEnd(size_t index)    { return m_config.m_end[index]; }
-        size_t LegLength(size_t index) { return abs(m_config.m_start[index] - m_config.m_end[index]) + 1; }
+        size_t LegLength(size_t index)
+        {
+            return abs((int)m_config.m_start[index] - (int)m_config.m_end[index]) + 1;
+        }
 
     public:
         void SetColor(CRGB rgb);
